add simple browser based user interface
Changes:

The `API` as well as the `TransferManager` now support simple callback handling.
Callbacks are intended to return a (callback_id, status) feedback of requested transfers
from the API on completion of a `MediatedTransferTask`.
This is only supported by mediated transfers, callback handling for direct transfers is still missing.

New Functionality:

A `WebUI` module was added, which supports transfer creation and transfer-status feedback.
The UI can be tested with running the `tests/test_webui.py` (instructions are in the file).
Backend (`web_ui.py` / gevent-websocket) and frontend (`raiden.js` / autobahn.js)
communicate using the WAMP-protocol (v.1), combining RPC-calls and PubSub-events.

To Do:
- implement PubSub messages in raiden base-code for further UI-functionality