moved events out of libraries into the contracts
this change assumes that if the library function didn't throw then the
execution was sucessful, in the future returning bool success might be
a better approach.

currently the transfer of asset is a dark area, the library is doing
the transfer but it should probably be done by the contract instead.