added RevealSecret message
The `Secret` message was being used to update the partner state and to
reveal the secret to the interest parties, added a new message for just
revealing a secret without the state update semantics.

To allow the `initiator` to check the receiving amount by the `target`,
`amount` was added into the `SecretRequest` message.

`asset` was added in the `Secret` message to force a different echo hash
for the `Secret` messages, important for exchanges were the `secret`
is shared but two assets are involved.

Splitted the method `claim_lock` in two (`withdraw_lock` and
`release_lock`). `release_lock` is used by the sending end of the
channel to update it's own channel, `withdraw_lock` is used by the
receiving end once a state update is received.

Removed hardcoded constantes for expiration timeouts (usually 5), using
`settle_timeout - reveal_timeout`.

Note: the message `RevealSecret` was added to the code base, but the
tasks are not using it yet.