untangle and cleanup raiden.mtree
The previous implementation of merkleroot was hard to understand, since
it tried to do several things at the same time.

I've tried to untangle that and now we've got separate functions to
- iterate pairwise over a list of elements
- compute the layers of the merkletree
- create proofs from the merkletree layers

I've also added a Merkletree class, which is initialized with a list of
hashes. This allows to compute multiple proofs for the same tree without
recomputing hashes all the time.

The merkleroot function still exists, but eventually it will be removed.

This is work-in-progress in order to resolve issue #202.