Create bumpversion configuration #270
This allows us to create proof of concept releases with just two commands:
On the master branch call

    ./prepare_poc_release.sh

Which will create a new PR. After the PR got merged, call:

    ./release.sh

This will create a tag and push it to the main repository.

As a convention I propose that we release all PoC-versions in the `0.0.x` series,
where `x` equals the PoC number.

Please note that due to the internal workings of `bumpversion` the PoC tags will
follow the `semver` standard, too (i.e. `v0.0.x` for `PoC-x`). Since pypi published
versions should already follow this standard, I believe this inconsistency with
our prior PoC releases is okay.