include API-exposure (http-requests), include push-events (socketio)
This is the server for the API exposure.
All methods marked by the @expose decorator will be available
via http-requests at the given REST endpoint,
once the Dispatcher is initialized and the server is running.
Synchronous methods will return their return-value immediately
in the request's respond message.
Asynchronous methods will not return their return value,
but rather a unique callback-id, that will later be used
in a 'rpc_response' Event via socket.io for evaluation.

The response is structured like this:

{
    callback_id: value or None if Synchronous
    method= str (the methods name)
    result= value or None if Asynchronous
}

Clients should also connect to the websocket
that is used for push-Notifications ('Events')
via the socket.io protocol.
Without the connection to the socket,
asynchronous calls over http-requests will be initiated,
but the client has no knowledge about the success of the call.
Also externally initiated state-changes such as channel-opening
or incoming transfers will not be observed by the client,
if he doesn't want to pull this information constantly.

Events are structured like this:

    Event:      'rpc_response'
    namespace:  '/'
    data:       {
                    callback_id: <int>,
                    procedure: <str>,
                    status: <bool>
                }

closes #10
related issue: #207