Use mappings to keep track of indices
This commit makes quite some changes to ChannelManager

Most importantly it changes the data structure of ChannelManagerLibrary
and ChannelManagerContract. We now use mappings to store the entries in
the library to avoid the problem in issue 232. Futhermore we use
mappings to keep track of the indices of the elements in the arrays in
ChannelManagerContract. This means that we don't have to iterate through
the arrays, thus solving the problem of potentially running out of gas
while iterating through the arrays.

This commit also introduces a rewrite of ChannelManagerContract and
ChannelManagerLibrary where the library no longer has any arrays.