Fix `test_transfer_from_outdated`
The test now uses the `wait_until_block` pattern instead of explicit
mining and sleeping.

However, there is a race condition with the settlement-task:

the second call to `set_settled()` will raise a `RuntimeError`, because the
channel's `_settled_block` will be updated between the precondition check
in `raiden/channel.py::blockalarm_for_settle` and the actual execution of
`set_settled`.

I added a second conditional to avoid this issue.