Removed Timeout message.
We need to assume that the next hop in the chain may be lying, in which
case the timeout message is useless.

For a mediator node, if the next hop is lying there is nothing he may
do, since it needs to wait for the lock expiration anyways, as for the

An initiator node should not depend on a downstream hop to decide when a
timeout happens, but instead use a local configuration for deciding when
a rety should occur.