Allow to store and use precompiled ABIs
This adds tooling to read the contract ABIs from a json file instead of
compiling them at runtime. It will always check the contract source files
against a recorded `sha1` checksum (and compile in case of a mismatch).

The interface for creating the precompiled files is a little funky though. For
precompiling, you should change `raiden.utils.STORE_PRECOMPILED` to `True` and
import `raiden.blockchain.abi`.

    from raiden import utils
    utils.STORE_PRECOMPILED = True
    from raiden.blockchain import abi
    # the import will already trigger compilation and saving

Since this is a packaging operation only, I believe this drawback is okay.

Fixes #428