Bootnodes and nodekeyhex seem to be unused in test
I don't see a reason to have bootnodes and as an extension nodekeyhex in
the test geth arguments.

If there is a reason then then at least the argument should be sanitized
to not be given without a value as I have seen it multiple times given
without a bootnode value in tests.

We are lucky geth accepts this at the moment, but this should not be our
default behaviour.