Check available memory before running the geth miner
If the system running the tests doesn't have enough memory to keep all
the DAG in the main memory the POW process will be slower than expected,
resulting in test flakiness, specifically for the tests that depend on
the confirmation of transactions, since this is a hard to diagnose
problem this commit adds an explicit check for the free memory and fails
the tests early. Note this will only fail the tests that are using the
geth miner, nothing happens with the purely unit/tester tests.