Removed transferredAmount from NettingChannelContract
The function transferredAmount was used to check if the courtesy
transaction was correct, since the courtesy transaction was removed
this function is not necessary.

Note: Additionally the transferred_amount is not sufficient to detect if
the latest valid transaction was provided, at least the nonce should be
checked but ideally the values `nonce`, `transferred_amount`, and
`locksroot` should be verified.