Default test timeout of 9min set globally
Some tests depend on concurrent actions that might not happen if a bug
is introduced, to detect this tests the test suite uses per test
timeouts to force failure. This strategy while may be good to control
the waiting time of a developer is not really portable, since each
machine will have different resources available and the same test may
run faster on one machine than another.

This is a problem specially with the build server and new relases of
third-party sofware, since our intent is to force failure for test that
would never finish in one hand, and in another to get a proper output
from the build server, it should be sufficient to have large timeout
that guarantees the later, and for the former a developer can set a flag
on the command line to overwrite the default.