Don't call `register_netting_channel` from `RaidenAPI`
Instead wait for the on-chain event to be processed before returning and
return the channel instance registered by the blockchain event handler
in `raiden_service`.

This comes at the cost of higher latency until returning, but leads to
better consistency of the `ChannelEndState` objects -- previously they
could be created before the `opened_block`, that controls the nonce
range, was set.