Preserving protocol queue items until ACK
- The queue must keep all the pending messages in-order for the
shutdown to save it's content.
- Moved constants from the Protocol class to the settings module
- Removed the special class NotifyingQueue, it was not required.
- Removed the global ttl chache, it was probably fine, but that state
was shared on test instances.
- Added a lock to protect the channel_queue dictionary, the current
code should be okay for greenlet based tasks, but the lock protects the
code for future thread models (echohash_asyncresult should also be
protected with a lock)
- Changed the semantics of the `_send_queued_messages` task, the prior
behavior simply dropped the current message once the retries were
exhausted, the new behavior will do a exponential backoff and only
proceed with the other messages once the current one is acknowledged
for.