Fix test usage of `detail()` call
There is no reason to allow for the address argument in `detail()`,
because it only mirrors the `our_*` and `partner_*` items in the result.
In the tests addressed here, I change the usage to make assertions on
`partner_balance` for the `detail(None)` of our channel instead of calling it
"backwards" and checking on `our_balance` on the partner's channel with
our_address set.