Ignore transfers after `leave` was called
This introduces a stronger concept of "leaving" a token network. Once
we're in the leaving state, the message handler won't ACK incoming
transfers (for the given token).

This allows to only close and settle all channels that have ever
received a transfer, without being susceptible to be stalled forever (by
receiving new transfers on newly opened channels).