xfail test for recovery from Write Ahead Log
This PR adds two xfail tests for recovering from the WAL

The first test tests that recovery with no un-applied state changes
works as expected.

The second test tests that recovery works even if a state change was
written to the WAL, but was not yet applied on the internal state of the
raiden node.