Do not invoke contract compilation at file import
Until now anything that imported `raiden/blockchain/abi.py` invoked the
compilation of our contracts in order to get the ABI and some other
attributes.

This PR moves the logic to a class that is initialized at import but not
instantiated there. The ABI and other contracts attributes are lazily
evaluated. In other words the first time something is required only then
are the contracts compiled.

Finally now we can do `raiden --help` without waiting 5-6 seconds.