Rename pregenerate_dag fixture for clarity
As
[this](https://github.com/raiden-network/raiden/pull/665#discussion_r126122831)
comment shows it is easy to mistake the functionality of the fixture due
to its name.

I assumed that the DAG is pregenerated once per session, while in fact
the fixture simply ensures it is there by generating it if it is not and
doing nothing if it is.

Renamed for clarity.