Accept messages with expired locks.
- A node may go offline while there are locked transfers in-transit /
waiting confirmation.
- If the node comes back online *after* any of these transfers have
expired, it *must* accept them to allow progress. This happens because
the partner/sender must order the messages to keep the messages
consistent, and the sender has no logic to recompute/removed outdated
messages from the queue. Without accepting outdated messages the
protocol won't make progress.

Note: Accepting a message with an outdated lock is only to progress the
queue, the transfer wont complete because of it, even if the lock is
learned afterwards, and the message wont be mediated because there
isn't enough blocks left to safely mediate a transfer.