Initial setup for using clique PoA in tests
Note: so far this only works for a subset of the tests, because there
are different miner accounts set up in the fixtures, but so far this
uses only one hardcoded address in the clique genesis setup.

To fix this, we need to pass the miner's account address to the genesis
generation and properly include it in the `extraData` field.