removed lock expiration check from netting channel
The netting channel is not at a position to properly check the
expiration of the lock. The expiration of the lock is defined by the
transfer task, and it takes into account the current pending number of
block until settlement, the payer lock.expiration, and the channel
reveal_timeout configuration. Once the function to create the locked
transfer is called the lock expiration was already decremented the
reveal timeout and the additional blocks are not required for safety.

An alternative to remove the check from the netting channel is to
provide the base expiration, and the create_lockedtransfer method then
decrements the reveal_timeout prior to creating the transfer, this
allows the channel and the transfer task that there are the required
number of blocks for safety.