Allow for delay when depositing
In the "partner-opened-first" scenario, where we hit a race condition
when creating a new channel, we want to deposit into the channel.
However, it may happen, that the channel exists on-chain, but our local
state is not yet up-to-date.
I added a timeout-bound waiting loop for our internal state to catch up
before trying to deposit.