Add txpool.pricebump geth argument
This is not a correct solution for issue
https://github.com/raiden-network/raiden/issues/804, of having the
`JSONRPCClientReplyError: replacement transaction underpriced` pop up in
tests, but at least it seems to reduce its occurence.

I believe it does not hurt to have this argument in the travis tests but
as #804 states we need to find out why in some occurences we are trying
to push for a transaction with the same nonce to our geth node.