Test recreating issue #784
In the issue one raiden node registers the token and the second tries to
register but fails with "token already registered" while its token list
is empty.

The Rest API call for token registration first checks if
`manager_address_if_token_registered()` and only if it's not proceeds
with registering.

If it is then it returns 409, but without having registered it with the
node and filled in the tokens list. This registering will happen later
if the "TokenAdded" blockchain event is processed but there is a window
where we can miss it.