Add our_state.nonce to serialization
This could be a possible solution to
https://github.com/raiden-network/raiden/issues/903.

If we have made direct transfers in a channel and then restart and use
that channel again either for direct or mediated transfers the nonce our
partner would expect would be different than the one we send since it
was not being serialized.

With this PR we serialize it.