inverted participant state's balance proof
The python code was storing the balance proof structure in the
participant state of the receiving end, this is switched in respect to
the smart contract code and makes it a bit harder to access the
transferred amount of a node since both state objects are required. This
commit switches around the balance proof property to fix these issues.