Remove noop exception handling, raise on failed transactions
Deposit will only wait for the event to be fired
`deposit` transaction will be ran synchronously, waiting for it to be
mined, so no need to wait long on the `deposit` API method.
wait_until is used only to give event handling greenlets the oportunity
to fire and process the events.
Tests have shown that, while most of the time we can expect the nonces
to be respected, and multiple transactions to be mined in order,
possibly in the same block, it's not reliable, specially on test
network. So, removing the 'no-wait' options