Ignore git-tag version on non tagged commits
Our version reporting is a little brittle for non-released installs from
git: If there is no tag (yet), but the intended version maintained in
`setup.py::version` is higher than the latest available tag, the version
description will be incorrect.

With this commit, we ignore the tag part on un-tagged git commits, which
allows for a better version resolution, especially during CI (i.e. PRs
for a version bump).