Add the Changelog
Proposing to use a changelog from here and on since we got our first
official release. We can easily use
[changelog-cli](https://pypi.python.org/pypi/changelog-cli/0.3.3) to add
stuff to the log and integrate it into our release process.

This also adheres to the [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) guidelines.

The proposal also is for every PR that makes a notable change to also have a
changelog entry added from here and on.