Add the changelog - v2
Proposing to use a changelog from here and on since we got our first
official release. We should integrate this into our release process

This PR uses `releases` for the changelog. It's kind of cumbersome and
not flexible in the way it interprets changelog entries. Each bullet
point must adhere to a specific format as seen in the [usage
docs](http://releases.readthedocs.io/en/latest/usage.html) or else the
sphinx generation of readthedocs will throw an exception.

The proposal also is for every PR that makes a notable change to also have a changelog entry added from here and on.