fixed healthcheck tests
Tests assume that transfers can be sent at the start of the test
function, this required adding a wait for the healthcheck in the fixture
setup. The problem is that tests that need to drop packets, health check
tests, couldnt finish setting up the fixtures if packets are dropped.

This commit changes the unreliable transport to /not/ drop packets by
default, allowing the health check to work on fixture setup and the test
to run.