Dont use all of gas limit in a transaction
Using the entire gas limit in a transaction some times results in it not getting
mined irrespective of the gas price since miners prefer smaller
transactions. Using 80% of the gas limit is working reliably from my tests.

Note: The proper gas limit solution here would be to set it from estimate gas
since estimate gas is now fixed.

This is just a step towards that to fix a problem being seen when using raiden
in ropsten at the moment.