Serializing StateChange/Event objects with ver 4
Changed from the highest version available (-1) to an explicit version
of the pickle protocol, this was done to avoid future problems related
to protocol compatibility.

Additionally, the version 4 of the protocol was choosen because it's the
highest version at the time of the commit, and because of the typing
module at least python3.5 is already a requirement for Raiden.