Fix race condition for event handling
The `RaidenAPI` calls `RaidenService.poll_blockchain_events` explicitely to
ensure that a transaction recently mined has all the events processed.
This is to update the node state and ensure the user of `RaidenAPI` will
see the expected effects applied.

The problem is that the `AlarmTask` can race with the transaction polling,
and *if* the event handler context-switches the effects may not be
applied once the `RaidenAPI` call returns. The introduced lock will
serialize all calls to `poll_blockchain_events`, forcing the `RaidenAPI`
to wait until the `AlarmTask` finish processing the events.

Note: This change will break if the events are *not* processed in the
`AlarmTask` greenlet.