moved message id generation up
the code for the channel and mediated transfer state machines should not
worry about the transport message id. this commit adds a separtion of
the internal send events and the send events emitted by the node state
machine, now the node is responsible to work as a middleware, get the
internal send event, and produce the correct external event with a
message id.