Added Delivered message
The Delivered message is used by the receiver to inform the sender that
the message has been saved. Used for transports like TCP and UDP, that
don't guarantee message persistence.

Most messages sent to another node must be processed, otherwise the
sender and reciver nodes will get out of sync. To guarantee that a
message is processed it must be persisted, otherwise under power faults
the message is lost, therefor the sender must keep the message around
until the receiver has confirmed the message is saved.

With a naive implementation, where messages are discarded by the sender
with just an acknowledgement instead of persistence, the recipient node
may fault before saving the message or its effects, and the message is
lost.