Add connection test middleware to imitate old behaviour
This adds a new `connection_test_middleware` that imitates the old
behaviour of the `rpccall_with_retry` behaviour. If the stop event is
set it raises and `RaidenShuttingDown` error, otherwise an
`EthNodeCommunicationError` is raised when there is no connection.

This also adds an additional check for exceptions during app startup. An
example for this is the stale check.