fixed regression
The variables SYSTEM_ERROR and NOT_ERROR are configured differently in
the raiden client and the test cases, this is required because the
raiden client is adopting is crash only, meaning the process should quit
on any exception that is unhandled, which is not useful for tests, since
we want to execute the whole test suite even if an exception is not
handled.

For this reason the variables must overwritten only for the tests run,
which is done by the fixture `dont_exit_pytest`, the top-level assigment
is a bad idea since just importing the module will change the
global state, and this could be done on a normal run, effectively
breaking the assumption that raiden is crash only.