Remove channel_identifier usage to search for channels
This identifier has a nonce on the blockchain, which makes it very hard
to be calculated only from the parameters given in the request.
As there should never exist more than 1 channel open for each tuple
(registry_address, token_address, own_address, partner_address), we can
just use it to search for the channels we want to interact with.
Channel_identifier is still returned in the request, for information,
but its value is ignored in the tests.
[ci integration]