Fixed unlock event handling
This PR is introducing a second filter for netting channel events. This
was necessary because with only one filter some events were missing.

The filter which missed events was setup with the following topics:

  [
     [None, channel_i],
     [ChannerlUnlock, participant1, participant2],
     [ChannerlUnlock, participant2, participant1],
  ]

With the above topic, the filter was *not* returning the event
ChannelSettled. The ChannelSettled event *was* returned with the
following topic:

  [[None, channel_id]]

Therefor, one of the instaleed filters use the topic above. In order to
get the Unlock events a second filter is installed with the following
topic:

  [
     [ChannerlUnlock, participant1, participant2],
     [ChannerlUnlock, participant2, participant1],
  ]

Using two filters was sufficient to get all events, but it is also
returning duplicated ChannelSettled events. The reason for this is
unknown, but the code is properly handling the duplicated events.

[ci integration]