Removed block_number from the events storage
The block number is provided through the
RaidenService.handle_state_change function, the block number used in
this function varies with the type of the state change.

Events from the blockchain can use a consistent value (the block number
from the event), but state changes for user actions or protocol messages
must use the current block number known by the node, which will vary on
restart. To avoid inconsistencies on the block_number on restarts, this
value was removed.

[ci integration]