Properly synchronizing the raiden node with the BC
During startup all filters must be queried to synchronize the node's
state. This must be done after the node's state is primed (either
initialized or restored from a snapshot), the filters are installed, and
prior to the transport being started.

This changes:
- `install_and_query_all_blockchain_filters` was using
`RaidenService.get_block_number` as the value for the filters
`to_block`, which is not updated until the alarm task is executed,
meaning the node's state was not updated before the AlarmTask was
scheduled.
- Removed from the `AlarmTask` an unused attribute named `response_queue`
- Removed the buggy sleep timeout computation from the `AlarmTask` (the
work_time was always zero)
- Inlined the method poll_for_new_block in the AlarmTask loop
- Added a check if the chain_id of the underlying ethereum node changes

[ci integration]