Added states changes to clear failed transactions
Not all transactions can be recovered, that is why the state changes
have a very explicit name with the prefix
`ContractReceiveRecoverableFailure`.

These state changes will have to be dispatched by the proxy layer, which
must detect when a given transaction failed for a non fatal reason.

These transactions are recoverable if the reason for the failure was a
race, and the local node lost the race to it's partner:

- BatchUnlock: The same locksroot must be registered
- ChannelClosed: The channel state must be at least closed
- ChannelSettled: The channel state must be at least settled

This transaction is recoverable if the reason for the failure is a race
with any other node:

- SecretReveal: The secret must be registered

This transaction is recoverable if the reason for failure is a race with
a monitoring service:

- UpdateTransfer: The nonce must be equal to the one sent by the raiden
node

[ci integration]