Fixed race for proxies instantiation
It's really important that each Raiden instance only used one proxy per
smart contract, otherwise the proxy synchronization wont work. The token
network proxy is an example, where the proxy needs locking to forbid
concurrent deposits (to avoid races with the ERC20 allowance).

To enforce a singleton proxy the blockchain service is used as a proxy
factory. The problem was that each proxy constructor checks the
blockchain data, which does IO and allows another greenlet to run, this
made it possible for races among two greenlets that need the same proxy.

Tho fix the above race coarse locks were added to protected the
instantiaton of each type of proxy.

[ci integration]