restapi: check the Raiden service is running
Because the RestAPI is a wrapper around the Raiden service, it's
possible to stop the service before the API. This will lead to
unexpected results, since the API will accept REST requests but the
requests cannot be completed.

This adds a check prior to every request, to make sure the service is
running. Ideally, the Rest API and Raiden Service lifecycle would be
managed togheter, but this commit does not introduce the required
changes for this.

[ci integration]