Remove flake8-coding module
It does not work with python 3.7 and renders flake8 unusable (it crashes). An issue along
with a fix has been reported to the maintainer
[here](https://github.com/tk0miya/flake8-coding/issues/16) but the repo seems to
no longer be actively maintained.

And to be honest we don't really need this module. The only thing it adds is a
check for the magic comment in the beginning of the file.

[skip ci]