Added RaidenEventHandler class
Made the event handler into a class and allowed it to be injected
through the App and RaidenService. This class can now be overwritten and
the protocol can be controlled by a custom handler class. This is
intended to be used in testing, where control over the messages is
required, e.g. a test for the secret being registered on-chain needs to
stop the unlock message.

[ci integration]