Decentralize discovery room
Now, a discovery room with canonical name is tried in all available
servers, starting with ours. If can join in a different server than
ours, but our server doesn't have an alias for it, create the alias, so
every server shares a single room, each one aliasing it to the canonical
name on itself.
A discovery room creating is attempted only if can't join in any
available server, and is only retried because of races in tests. On real
usage, it shouldn't be a problem, and keeps the room decentralized
anyway
[ci integration]