REST Api: removed host/port from start method
The `start` method had two arguments, host and port. This has multiple
problems:

1. It breaks the interface, meaning other runnabled work with `stop`
followed by `start`, but the rest api needs special logic to give to
`start` the host and port arguments.
2. It's easy to mess up if one forgets to pass around the host port pair
with the intance the restart will end up with server in the wrong port.

From above, 2. the reason for some flakyness in the crash stress test,
where part of the test is to force failure in the REST API by restarting
it, but without the configuration the server changed ports and ended up
causing port conflicts (when two servers were restarted and both tried
to use the default port)