removed module level loggers
structlog's are lazy, and will be configured on first usage, this allows
the module level loggers to be instantiated at import time and be
configured latter. Although this is convenient, it doesn't allow for the
loggers to be reconfigured during a test session, since the logger will
stick to the first configuration used in the session, and further calls
to configure will not have effect since a new logger is not
instantiated.

This fixes the above, and allows tests to reconfigure the the logging
library. This is a step towards reducing the log size in the build,
which become necessary since travis has a maximum size for logs.

[ci integration]