Simplified the filter polling
The default implementation of filters provided by web3.py rely on the
ethereum node to keep the state of the from_block number. Although this
is convenient it's a problem when either the Ethereum or Raiden node are
restarted, because the previous block number must be remembered to
re-install the filter.

Since the from_block needs to be known anyways instead of relying on the
remote ethereum node to know the from_block the stateless filter was
introduced, which always rely on the start_block number being known.

This just removes the unused code which supported the web3.py Filter
class.

[ci integration]