REST API: tracking all greenlets spawned by the REST server
By default the WSGI server will fire and forget a greenlet to handle
incoming data. This means during shuting down the WSGI server cannot
clear the started greenlets. The Group can be used to track and do a
clean shutdown.

The same applied to the UDPTransport, it was not tracking all the
greenlets spawned to handle network packets, which led to unclean stops
and orphan greenlets.