events: only poll confirmed events
Reorgs can introduce duplicated events in the system, and it has been
experienced in the past with a close transaction. To handle reorgs there
are two options:

- Raiden should not poll for events from the latest block, but only from
latest - confirmation blocks.
- Raiden keeps in sync with latest, and if necessary also fork locally
by replaying the state changes.

The first option has the adavantage of being simpler to implement, while
the second allows Raiden to react faster to some events, e.g. secret
being registered in the registry.

This commit implements the first.