bugfix: fixtures must wait for network creation
Without waiting for the number of confirmation blocks there was enough
time during the channel setup to update the node's state. With the
confirmation blocks, however, it's requied to wait for the token network
to be available. This is specially true for test that create multiple
nodes but no channels.