Added checks for update transfer proxy
- Moved the check for the channel being closed before the transaction
is sent, if the channel is not closed then the transaction will always
fail.
- Added a check after the transaction is mined for the settled state of
the channel, it's possible (but highly unlikely due to the settlement
window configuration) to have a race condition among the update transfer
and another settle transaction. Althought this is a really serious
configuration issue it's was classified as a recoverable error since the
problem may appear on badly configured nodes.
- Added a check for the signature of the channel partner.
- Improved the handling of exceptions when recovering the address from
the balance proof signature.