Added protocol version to the Ping message
The protocol version handshake should be tied to the healthcheck in
order to avoid race conditions. Where a target is known to be online but
its protocol version is unknown.

Without the version in the Ping message the sender must assume the
lowest protocol version, otherwise a message unknown o the target could
be sent. Because the new message is not known to the target, it would
never be processed, and the message queue for that channel would stall.

Adding this field to the Ping message fixes the above problem. This is
being added now just to introduce the strategy of protocol handshake for
future releases.