Added MessageHandler class
Adding a class to provide late binding, the previous code relied on the
free standing functions from the message_handler module to handle
messages. Although this worked it didn't provide for a mechanism to
extend the message handling, which is desirable for testing (waiting for
a message to be received before proceeding).

Note: an alternative would to just use the module as an object instead
of introducing the class, this however makes extending the behavior a
bit more burdensome, since alternative implementations would have to
reimplement the dispatching code.