Bind TokenNetworkGraphState to a token network
It turns out that the `TokenNetworkGraphState` currently doesn’t know
which token network it belongs to. We therefore need a breaking change
and add the token network identifier to the `TokenNetworkGraphState`.
With this the objects don’t compare to equal any more and proper
instances for each token network are created during deserialization.