test: remove expired lock must update state
The target must update the balance proof and merkletree once a remove
expired lock message is received, a test was added for channel handler,
which is used by the mediator and target state machines, asserting that
the values are set if the message is valid.

If a node fails to assing the balance proof to its local state the nodes
will desynchronize because the nonces will not match. Additionally, it
may lead unexpected behavior for on-chain transactions (the old balance
proof would be used, which would not lead to token loss, but is
unexpected behavior).