Add Supervisor class to Runnable interface
The Supervisor runnable allows to monitor a group of Runnable's in order
to kill the whole group in case of an exception in any of the monitored
children. This is meant to help with the `pytest-timeout` behavior in
tests with multiple app instances. (Depending on the moment when
`Timeout` was raised, it could happen in the past, that there were runaway
App instances that did not get killed).