bugfix: Mediator task must wait for expired lock
The mediator task definition of finalized was wrong. The task was
considered final the lock expired, which happens once the corresponding
block was mined.

The consequence is that during a run where the lock expired, the
mediator would send the remove expire lock message, and set the status
of the task as finalized, leading the task to be cleared. The problem is
that the mediator task is also responsible to handle the remove expired
lock messages, and if is clered these messages are not handled, leading
to synchronization problems among the participants.