bugfix: Proper life cycle for the mediator
The mediator task must be kept around while the mediated lock is present
in any of the used channels. This is necessary because eventually either
the lock will be unlocked or expire, and the corresponding messages
which remove the lock from the corresponding merkle tree must be
handled.

The code was ignoring this rule for a transfer which could *not* be
forwarded. The corresponding channel which received the message was
being forgotten, and the task cleared when the other channels where
properly finalized.

A mediator would fail to forward the transfer if there where no usable
routes. A route would be considered unusable for a number of different
reasons:

- The channel was closed on-chain
- The node didn't have enough capacity to forward the transfer at the
moment the lock was received
- The partner node was considered offline at the moment the transfer was
received
- The channels which didn't fit the above criteria where tried already.

This added the received transfer, which could not be forwarded, as a
waiting transfer, and the clearing of the task is now also checking for
the status of the lock in the channel which received the waiting
transfer.