Try to handle estimateGas predicting that a call will fail
After the changes introduced by
https://github.com/raiden-network/raiden/pull/2623/ we again rely on gas
estimation for the gas limit to provide to an onchain call which is great since
it reduces our overall ETH balance requirements.

But we don't take into account the fact that if estimateGas returns None it's
because the transaction will fail, either due to our account not having enough
funds or because the call will just revert somewhere down the line due to the
contract logic (e.g. calling settle before closing).

With that in mind we can now detect those bad cases a bit earlier and handle
them without sending an onchain transaction. This is what this PR is trying to achieve.