Bugfix: Before settle the locksroot is unknown
The code which calls unlocks used to search for the balance proof
through the locksroot value, this works for unlock, however it does not
work for calls to settle, since at that point in time the smart contract
only knowns about the balance_hash.

Instead of adding two new functions which do pretty much the same thing,
I decided to rely on the fact that balance_hash is collision resistent
and it contains the locksroot, meaning that a search for either value
should work, and as of version v0.8.0 of the smart contracts the balance
hash is available for when both function calls are done.