Fix deposit amount decreased crash
In development mode we unfortunately seem to still have limits as we
can see from the code
https://github.com/raiden-network/raiden-contracts/blob/7105cf818fc290107ad48df9497bddcdec5c540e/raiden_contracts/contracts/TokenNetwork.sol
of what is probably the pre_limits contracts.

The fix for now, is to also take that limit into account at
development mode and not allow the client to send a transaction when
the token amount is above the limit that development mode contracts
have.

Fixes #3135