Mypy on blockchain, encoding, storage, network
Four new directories are covered by mypy checks.

The most significant change is the introduction of a new type
``UDPMessageID`` for udp_utils.py.  This was necessary because
``MessagerID`` was used for very different two types (one was ``int``
but the other was a tuple).

Besides, some type annotations were adjusted as usual.