Fix CircleCI timeout with parallelism=1 by increasing to 2
CircleCI parallelism was decreased to 1 on
f34ab26bd5dbd4989dffdb1a134c909ee1e5c0b6 to avoid hitting the limit of 4
on the smaller plan and allow all 4 sets of integration tests running
together, but it pushed the time required to run a single set of tests
to more than 1 hour, which made it timeout and not complete successfuly
even if no test error reported fails.
This PR increase it to 2, to decrease the time of running a single set
of test to less than 1hr, but preventing all 4 sets of tests of running
parallely, running 2 then 2 at a time.