If the alarm task is restarted reset stop event
During testing we `stop()` and `start()` raiden again. This will also
stop and start the alarm task. After the initial stop of the alarm
task the `stop_event` was never set to False again so it never ran
after a restart.

This may be a cause for other failing integration/stress tests