Change batch_unlock on settlement logic
WIP note: this squashed commit needs to be broken down

This changes the unlock logic during settlement.
We now calculate the gain for either channel side to unlock

The new logic is:
If we have any sent transfers with expired locks, we will unlock the
sending side.
If we have any received transfers with on-chain revealed secrets, we
will unlock the receiving side.