removed deep copy from the state manager
This is a really important change.

The copy was introduced for two important reasons:

1. To treat the state as immutable.
2. All changes to the business logic must go through the WAL
   abstraction.

Regarding 1:

Updating state in-place destroys data by overwritting it, this is a
problem for equality reasoning and referential transparency. This is a
bit harder with python since the standard library doesn't have first
class immutable data structures, structural sharing has a high cost in
terms of code, and usage of data structure libraries were not a
priority.

To preserve referential transparency at the state tree level the copy
was used.

Regarding 2:

If the application code got a reference to the state object and directly
modified it without a state change, restarts would not work because
replaying state changes would not result in the same end state,
effectively corrupting the state machine's state.

*without the copy* The biggest problem with this is that on the first
run things would seem to work, because the corrupted state would be
shared among the application and the state machine. And this can easily
go unnoticed unit/integration tests that don't exercise the restart of
the node.

To prevent the application code from holding a reference to these
objects the copy was used.

The reasons above are important and probably helped prevent a category
of bugs, however the copy greatly increases the time necessary to apply
a single state change, so this commit reverts the usage of the deep
copy.

To move forward both discipline and a library for immutable data
structures should be used.