Additional details in logging for debugging
Issue #3196 proved to be hard to debug, part of the problem is that not
all data used to query the database is available, so it's hard to know
why the database record was not found. This adds the two missing data
fields to the log message, in case the error manifest itself again in
the future.