Python API skipping WAL to close channels
The work to offload waiting for transaction completion decoupled the
caller from the completion of a transactoin, this means that errors
which are informed through exceptions are raised on a different
location. This is problemation for the python api, since that
abstraction is used by the rest api, and without the exception there is
no simple way to inform the api caller what happened.

This sacrifices restartability of the close action in favor of better
error reporting.