fix unlock calls
The token network smart contract has a few optimizations, one of these
optimizations is to remove storage data as soon as possible, to reduce
gas usage.

Because of that, once the channel is settled the channel structs and
their corresponding participant structs are removed. This removes the
following data from the smart contract:

participant.deposit
participant.withdraw
participant.is_the_closer
participant.balance_hash
participant.nonce

and

channel.settle_block_number
channel.state

Meaning the processing of unlock requests cannot rely on the
availability of the balance_hash and nonce data to find the
corresponding merkle tree. For a proper unlock the locksroot must be
used.

In the other hand, because of privacy concerns, the data provided to the
functions closed and updateTransfer is blinded, this is done to avoid
sharing the all the details with the monitoring service. Therefore the
event which request the call to settle, which has to find the complete
balance proof can not rely on the locksroot for the search, it must use
the balance_hash.

Because of the above, now we have four functions to search for the
channel data, two for our data, two for the partner data, one of each
for the call to settle and unlock.