Properly handle SecretRequests for settled channels
When receiving a secret request for a settle channel there was no
check if the channel is settled (in which case the channel identifier
is removed from the mapping) and we were gonna get a KeyError.

This commit fixes that by properly handling the absence of the channel
identifier in the case of a SecretRequest for a settled channel and in
some other cases in the intiator code.