Workaround for parity pending block + call bug
In parity doing any call() with the 'pending' block no longer falls back
to the latest if no pending block is found but throws a mistaken error.
Until that bug is fixed we need to enforce special behaviour for parity
and use the latest block for checking.

Bug link: https://github.com/paritytech/parity-ethereum/issues/9707