Separate methods to handle state changes
A Raiden node will be driven by two major drivers, the transport layer
and the rest api.

ATM tese two drivers have conflicting requirements on how to handle the
complete state change cycle. The Matrix transport layer does not want to
have exceptions raised, nor does it want to block and wait until a state
change is completely processed. The rest api OTOH wants to wait for the
completion of a given state change, and if an error happens it has to
catch it to properly report it to the user application.

For this reason a new code path was introduced, which exposes the errors
from the processing of transactions. Namely the processing of the close
transaction, which can throw and the message of the exception must be
exposed to the API caller.