Fix high CPU usage when idle
Fix #3369

The commit which introduced this problem was
7e6dd0fd11aca00871556e1cdfa926fb2de415fa

It tried to fix the restart tests by resetting the alarm task's
`_stop_event` to `False`. Whenever a new alarm task is started during
tests.

The problem with this approach is that setting an `AsyncResult()` to
`False` at the very beginning, gives it a value. So by having a value
here:

https://github.com/raiden-network/raiden/blob/e76687c5d4dc8c6dd8888719ea60243f37c9a945/raiden/tasks.py#L151-L158

`self._stop_event.wait(sleep_time)` does not actually do any
sleeping. That's because the `AsyncResult` already has a value as we
can also see in the documentation
[here](http://www.gevent.org/api/gevent.event.html#gevent.event.AsyncResult.wait).

> If this instance already holds a value, it is returned immediately. If this instance already holds an exception, None is returned immediately.