bugfix: flaky test_matrix_cross_server_with_load_balance
The test was waiting for a single message in the wait loop, however it
was asserting that two messages arrived, this caused flakyness.

This additionally remove the asserts, since the number of messages is
already tests as part of the wait loop, which will timeout if the
expected number of messages is not received.