fix broken test and token network proxy unlock UnboundLocalError
This is a simple fix, which was prompted by me spotting that
`test_settled_lock` had a check for "generic" exception when giving
a lock that should prompt failure. The unlock call had the wrong
number of arguments but this was never detected due to the catch all
exception. As a result we had no test checking the unhappy path of
proxy's unlock function and we had an UnboundLocalError there.