fixed partial initialization
The contract manager is *required* to instantiate any of the proxies
through the BlockchainService. Without a contract manager the service is
effectively useless, since it's main responsability is to be a container
of proxies, to ensure proxy instances are shared making sure the locks
work.

This changes code to alwas give a reference to the contract manager
while instantiating a blockchain service.