Only check ethereum network id periodically
This PR takes the conversation from
[here](https://github.com/raiden-network/raiden/pull/3418#discussion_r254208431)
and presents an alternative approach.

Instead of having a cache on the network ID, we simply store it once
in the blockchain service at startup. From that point on we
periodically check the underlying ethereum node's network id and if it
changes under our feet we crash and burn (with a nice explanation).