removed race condition
The attribute `exception` is set by the greenlet itself, when an
exception is raised until the root level of its stack. The problem was
that the sleep around the time for the alarm task and the exception
handling have no synchronization what so ever, leaving the code to
timming conditions, depending on when each greenlet is scheduled. This
just waits on the greenlet it self, once the greenlet exists we know
that it will be because of the excpetion raised by our code, and then
the test can assert on the exception attribute.