Take any transfer rather than the initial one.
PaymentStatus needs to be registered with data
related to the transfer regardless of which one.
What this means is that the initial transfer with
the task's secrethash might have been removed;
Because it's lock expired or the secret is known.
But since the payment status needs to only
know about target, payment_identifier and
balance_proof, any transfer in the list should
be fine to use to register the payment status.