Feature: secret and/or hash with payment request
This feature allows the payer to optionally provide a secret and/or a secret_hash that should be used for the payment. If no secret and secret_hash provided as part of the payment request the secret and the hash should be selected by Raiden. If only a secret is provided the secret hash should be extracted from the given secret. If both secret and secret hash provided they should match. Providing only the secret hash is not currently supported.