Populate all ContractSend events with the blockhash at init
There is a problem that this commits makes a temporary hack around.

For `ContractSendChannelClose` and for `ContractSendSecretReveal` only
in the case a mediator receives an offchain secret reveal we don't
have any way of getting the block hash. We have the block number, but
we can't make an I/O RPC call to get the blockhash from the block
number.

For these cases I temporarily give a blockhash of bytes(0). It
probably does not need to be checked when the event is processed. The
proper solution here would be to always have the block hash available
by adding it to the chain state along with the block number.