Cleanup message class hierarchy
PR #3374 introduced a potential DoS vector: since the message specified
a command-ID, it could be sent to a client (instead of a service).

Since clients expect a `message_identifier` for protocol messages, this
would lead to `AttributeError` on processing e.g. a `RequestMonitoring`
message in the client.

A couple of different traits of messages (has `message_identifier`, has
`sender`, has `signature`), so far were rather implicit. I introduced a
couple of message classes, to allow for explicit definition _and_ allow
re-use of `SignedMessage` in non Raiden-Off-Chain-Protocol messages
(i.e. `RequestMonitoring` or `UpdatePFS`, #3477).