Geth: Moved logging to a separate folder
Currently a single test session will run multiple tests, each test
starts a new set of geth nodes /sharing/ the log paths, because geth
overwrites logs on subsequent runs this results in loss of logs.

This commit adds an explicit logpath to the fixtures, allowing all logs
to be collected in a single folder, which can be used to publish these
logs as artificats from the CI.

Because each new geth cluster uses a different set of private keys, the
address of each node will be different, meaning that log file names
won't collide.