Fix test race condition
The test `test_locked_transfer_secret_registered_onchain` was used to
guarantee that a node does not reuse a known secret. It first registered
the secret and then called `start_mediated_transfer_with_secret`. Because
the Raiden node was only querying the blockchain state from confirmed
blocks the recently registered secret was missed.

In this particular instance, using `latest` as a block_identifier is
fine, we don't need a consistent view of the blockchain for that
particular call.