Waiting for all secret transactions to be mined
If the user code is calling `register_secret` or `register_secret_batch`
it most likely expects the secret to be registered once the method
returns and if something goes wrong it probably expects an error. This,
however, was not the previous behavior, if there was an in-flight
transaction to register the secret the method just returned, regardless
of the status of the transaction and how long it would take for it to be
mined.