Feature: Allow inject secret-hash without a secret
With this commit the REST api allows the caller to provide a secret-hash without a secret. The intention is that the secret will be provided by the payee(target).
Since the target does not yet support this capability, in the case that only secret-hash provided, the code ignores the secret-hash, select a secret an use it. This will be removed next when the state machine is enhanced to support this mode.
The EventPaymentsSecretSuccess includes now the secret so that the API can should it to the user. The secret and secrethash removed from payment_status.