Introduce CanonicalIdentifier
As raised in #3493 we want to tighten the definition of what identifies
a channel.
This commit introduces the new type `CanonicalIdentifier` as a (named)
tuple of `(chain_identifier, token_network_address,
channel_identifier)`.

- In order to allow for a gradual introduction, I chose to use the
identifier with signed balance proofs first.
- In order to avoid migrations at this point in time, the identifier is
used as an `__init__` argument, but then destructured into the existing
fields.
- To minimize the amount of mypy casts introduced by this,
`CanonicalIdentifier.token_network_address` is typed as a union of
`TokenNetworkAddress` and `TokenNetworkID` for now.