Reorginized initialization code
The code was split into side effect free, and effectful. The side effect
free code is named using the prefix `_initialize` as a convention, these
functions are preparing the node's state for proper operation. Effectful
functions are named with the prefix `_start` and start a subservice,
namely the alarm task or the transport layer, which are effectively
applying state from the outside world. The order in which these
functions have to be called is partially described by the use of asserts
that check for the expected state.

Note that the call to `first_run` is not effectful because the flag
`ready_to_process_events` is set to `False`.