Fixed skipif mark
Calling skipif from inside a fixture means that pytest has to execute
the fixture in order to know if the test should be executed or not. It
happens to be the case that fixtures are executed in the same order as
the test function parameter list, meaning that if the skip_if fixture is
at the end of the list, time will be wasted executing the fixtures
before it.

This removes the skip_if fixtures, and uses the pytest.mark.skip_if
instead, allowing the fixtures to be exited early.