Refactor ReleaseKeeper Class.
This refactoring aims to split the existing code
up logically, providing two new
classes, as well a multiple new methods for
managing releases.

The ReleaseArchive class provides a more pythonic
approach to reading and extracting a release's
archive file (.tar.gz or .zip).

The Release class now takes care exclusively of
downloading, unppacking, installing, uninstalling
and removing a single release to and from the
local machine.

The ReleaseManager manages several Release instances,
and supplies convenience methods to install
and uninstall a release, as well as a purge method to
uninstall and remove all releases.

Finally, we moved these classes to a new file, as
we feel their complexity warrants a dedicated module.