Introduce placeholder for chain_identifier
In scope of #3493, there are a number of classes that shall be
identified by a `CanonicalIdentifier`, but do not yet serialize the
chain identifier.

In order to being able to introduce `canonical_identifer` as a
constructor argument for those, I define a placeholder "chain
identifier", that allows me to proceed with the expand phase of the
refactoring, without breaking the (de-)serialization scheme yet.