Use `CanonicalIdentifier` in `ContractReceiveChannelBatchUnlock` constructor
Note: this is one of the slightly more questionable changes:
`ContractReceiveChannelBatchUnlock` is created from a blockchain event,
that does **not** specify the channel identifier (because it is created
_after_ settlement and hence the channel data does not exist in the
contract anymore. However, while processing we will recover it (if the
event is applicable), so in my opinion, we could recover it in
`raiden/blockchain_events_handler.py::handle_batch_unlock`, and treat
the event as a fully identified event.