Tightened version check.
A upgrade may fail and leave a database with an old version behind,
however, it should never be the case that the current database has a
higher version number then the expected. Additionally, the existence of
the settings table with a version entry is necessary, otherwise the db
is corrupted.