Bugfix: Handle process kill during migrations
This commit splits up the cleaning of partially upgrade databases and
the selection of the latest database. The reason is that a previous
commit changed the semantics of the function `older_db_file`, which was
suppose to return the largest database version, which is smaller then
the current one. By splitting the two steps it is not necessary anymore
to keep track of the old behavior of the function.