Parity between solc version used in client and contracts
We use solc in raiden only to compile one contract in the assumptions
test.

Our tests right now don't work with solidity >= v0.5.0 and the
contracts actually use solidity v0.5.4.

This PR fixes the tests and makes them work with solidity v0.5.4 which
is the same as in the contracts repo.