Fix precedence for CLI options / config file settings
Previously precedence between options given on the command line and set
in the config file was established by comparing the current value to
the option's default.

This breaks when the config file specifies a value that is different
from the default and the user then wanted to restore the value to the
default via the command line. In this case the cli option would be
silently ignored since it was impossible to distinguish the explicitly
given value from the baked-in default.