Feature: Allow inject secret-hash without a secret
The payee can use an external secret resolver to resolve the secret hash to the secret instead of asking the payer to reveal the secret. This mechanism will be used only if resolver-endpoint configured by the payee. If a configured resolver failed to resolve the hash to the secret the payee will send a "normal" reveal secret request back to the payer.

Main changes:
1. CLI - Hash resolver options group. resolver-endpoint option. Using this option the payee can set  the URL of the resolver that will be used to resolve the hash to the secret.
2. CLI allows to provide a secret-hash without providing the secret.
3. resolver/client - is used to send an http/https request to the resolver while providing in the request body a json with data needed by the resolver
3. resolver/dummy_server - can be used to test the mechanism without full integration with fully functional resolver