get_random_service() should use the confirmed block
The solution is not ideal as it's not using Raiden Chain State's
confirmed block, but the two should be identical.

The reason why this is not using Raiden chain state's confirmed block
is that at the startup when the proxies are instantiated is when
`get_random_service()` is called. And at that point there is no raiden
App and not raiden service so the chain state has not been initialized yet.