Fix message to PFS
The JSON message must hex encode all byte-typed values. This was not the
case for
 * from
 * to
 * signature

Also move the IOU into it's own object as expected by the PFS instead of
adding all keys to the outermost JSON object.

The PFS request now declares the POST content as `application/json`
content.