Move CI PyPI deployment after binary building
In the past we had multiple cases where building the binaries on CI
failed for one reason or another.
This caused a problem since the PyPI upload used to happen in parallel
with building the binaries and therefore could happen despite the
binaries failed to build.

Additionally PyPI doesn't allow to overwrite a release, which then causes
problems in subsequent runs fixing the binary building.

[skip tests]