Clear ContractSendChannelBatchUnlock on restart properly
This solves the problem of having a Batch Unlock transaction in the
pending transaction queue which doesn't get cleared properly.

The pending transaction doesn't get cleared because our transaction
invalidation logic includes checks for the transaction clearing our
partner's side of the channel which is not true any more since the
current node could be trying to unlock it's own side of the channel.

The change basically clears batch unlock pending transaction in case
both sides of the channel were unlocked (channel got cleared).