tests: fixed race condition
test_recovery_happy_case asserts on the state of mediator channels, but
the existing tools to do transfers do not synchronize the mediators.
This adds a new tool that will assume the provided path is used and wait
until all nodes are synchronized.

note: the tool can be improved by checking the locked transfer message
is sent to the exptected node, this however can be done as a later step.