Add migration test for v21->v22 [DEV version]
This adds a unit test for the migration from db version 21 to 22. This
is the development version, which runs the test against _all_ db
serializations found in a folder `/tmp/dings`, in the format they are
created from running `pytest raiden/tests/integration` from the branch
https://github.com/konradkonrad/raiden/tree/create_migrations_data
(rebased on latest master).