Fixed secret registration
The block identifier is unecessary for the secret registration because
the action is always valid (there are no preconditions to check), for
this reason the blockhash is unecessary.

In order to preserve a consistent blockchain view, this removed the
unecessary argument and start using the latest confirmed block as a
reference point. This removes problems with prunned blocks and chain
reorgs.