Silence mypy
Before #3863 we will still query path-finding for a route for a refund
transfer. We are not supposed to query for routes with a
`PaymentAmountWithFee`, but only with a bare `PaymentAmount`. However,
such is not available at this point.
Because permissive source routing will likely abolish the need for route
querying at this point, I opted to just silence the error, by an
(incorrect) type cast.