Fixed data races on node state
The event handler for a ContractSendChannelBatchUnlock needs to load the
channel's data from the ChainState. However the event handler and the
state change dispatch are /not/ synchronized, meaning that once the
event handler has a chance to run other events may have been processed
meanwhile.

Consider this:

- Node A process the contract receive channel settled event, as a
  side-effect it produces a contract send batch unlock to unlock the
  only pending merkle tree (say from A->B). This event is then added to
  a queue of raiden events to be processed (currently the queue is not
  explicit, but it exisits in the form of multiple greenlets waiting to
  be executed). This means the processing of the event may take a while.
- At the same time Node B process the same event, and it too decides to
  unlock the merkle tree.
- Now, assume that node A is under heavy cpu load, whereas node B has
  plenty of computational resources available, additionally B has an
  agreesive gas price setting, used to mine transactions faster. If it
  takes B `n` blocks to process the settle event, send the unlock
  transaction, and for the transaction to be mined, this is the maximum
  amount of time that A may take to process its send batch unlock.
  Otherwise once A starts processing the send batch unlock, the channel
  may have been removed already.

This may be a bit unlikely, but it's a possible race condition. Restarts
don't have a problem with this because the batch unlock would be
cleared from the pending transaction queue during start up and it would
never be sent.