Reintroduced psutils based port generator
psutils cannot be used for mac, however it works reliably in the CI
system. The solution based on socket bind/TIME_WAIT is flaky and leads
to multiple build failures a day. This adds back the psutils for
non-macos systems only, the goal is to decrease the number of flaky
build failures.

Note: The problem may persist for MacOS systems, since SO_REUSEADDR
behaves differently across Unixes, ref:

https://stackoverflow.com/questions/14388706/socket-options-so-reuseaddr-and-so-reuseport-how-do-they-differ-do-they-mean-t