unlock handling state pruning
Handle pruned data on channel settle event

Fixes #3856 . The handler for the channel settle blockchain event may be
executed long after the event happened. This may lead to errors because
the node has to query the blockchain to recover th locksroot, and by
default it uses the block at which the event was emitted.

This adds a exception handler which just retries the query using the
latest block. The rationale is that it is preferrable to have the
contract receive channel settle with the correct data, which is
available at the block in which settle was called. However, there is no
risk of loss of funds if the data is not recovered, so in order to
continue the node's operation if the emitting block was pruned, then a
new query is done using latest.