Consistency changes.
The smart contract uses the arguments participant and partner
differently for the function `getChannelParticipantInfo` and `unlock`.
This changes the proxy to use a consistent order of argument in the
Python, the idea is to reduce confusion.

On future versions of the smart contracts the parameters will be fixed,
probably by renaming.