Fixed argument order and simplified names
The arguments order for the function getChannelParticipantInfo and
unlock are swapped around, this led to a simple but time consuming bug
to figured out. This fixes the problem and uses new variables names to
clarify the direction of the balance proofs (not the direction of the
tokens after unlock, because that could go in either direction).